+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 *Also see [Tools and installer changelog](https://github.com/clojure/brew-install/blob/1.10.3/CHANGELOG.md)*
 
 * next
+  * TDEPS-213 Add `-X:deps aliases` to list available aliases
   * TDEPS-226 More nunanced error handling for s3 downloads
   * Better error message when git url can't be inferred
 * 0.14.1178 on Apr 22, 2022